+++ include/pc.h
@@ -32,10 +32,10 @@ void		inportsw (unsigned short _port, unsigned short *_buf, unsigned _len);
 void		inportsl (unsigned short _port, uint32_t      *_buf, unsigned _len);
 void		outportb (unsigned short _port, unsigned char  _data);
 void		outportw (unsigned short _port, unsigned short _data);
-void		outportl (unsigned short _port, unsigned long  _data);
+void		outportl (unsigned short _port, uint32_t _data);
 void		outportsb(unsigned short _port, const unsigned char  *_buf, unsigned _len);
 void		outportsw(unsigned short _port, const unsigned short *_buf, unsigned _len);
-void		outportsl(unsigned short _port, const unsigned long  *_buf, unsigned _len);
+void		outportsl(unsigned short _port, const uint32_t *_buf, unsigned _len);
 
 unsigned char	inp(unsigned short _port);
 unsigned short	inpw(unsigned short _port);
